# graphite-rest/openapi.json
{
  "components": {
    "schemas": {
...
      "ArtifactOut": {
        "properties": {
          "artifact_type": {
            "title": "Artifact Type",
            "type": "string"
          },
          "citations": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/GraphEvidenceOut"
            },
            "title": "Citations",
            "type": "array"
          },
          "content": {
            "additionalProperties": true,
            "title": "Content",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "study_step_id": {
            "format": "uuid",
            "title": "Study Step Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "study_step_id",
          "artifact_type",
          "content",
          "created_at"
        ],
        "title": "ArtifactOut",
        "type": "object"
      },
...
      "GraphEvidenceOut": {
        "properties": {
          "chunk_id": {
            "format": "uuid",
            "title": "Chunk Id",
            "type": "string"
          },
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "excerpt": {
            "title": "Excerpt",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "chunk_id",
          "document_id",
          "label",
          "excerpt"
        ],
        "title": "GraphEvidenceOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "graphite API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/narrations/{narration_id}/audio": {
      "get": {
        "operationId": "narration_audio_api_v1_narrations__narration_id__audio_get",
        "parameters": [
          {
            "in": "path",
            "name": "narration_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Narration Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Narration Audio",
        "tags": [
          "narration"
        ]
      }
    },
    "/api/v1/study-artifacts/{artifact_id}/narration": {
      "post": {
        "operationId": "create_narration_api_v1_study_artifacts__artifact_id__narration_post",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Narration",
        "tags": [
          "narration"
        ]
      }
    },
...
  }
}